Crate google_proximitybeacon1_beta1

This documentation was generated from proximitybeacon crate version 0.1.15+20160429, where 20160429 is the exact revision of the proximitybeacon:v1beta1 schema built by the mako code generator v0.1.15.

Everything else about the proximitybeacon v1_beta1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

beaconinfo

getforobserved

beacons

activate, attachments batch delete, attachments create, attachments delete, attachments list, deactivate, decommission, diagnostics list, get, list, register and update

namespaces

list and update

Other activities are ...

get eidparams

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.beacons().attachments_list(...).doit()
let r = hub.beacons().get(...).doit()
let r = hub.beacons().attachments_create(...).doit()
let r = hub.beacons().decommission(...).doit()
let r = hub.beacons().activate(...).doit()
let r = hub.beacons().list(...).doit()
let r = hub.beacons().update(...).doit()
let r = hub.beacons().attachments_delete(...).doit()
let r = hub.beacons().deactivate(...).doit()
let r = hub.beacons().register(...).doit()
let r = hub.beacons().diagnostics_list(...).doit()
let r = hub.beacons().attachments_batch_delete(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-proximitybeacon1_beta1 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_proximitybeacon1_beta1 as proximitybeacon1_beta1;
use proximitybeacon1_beta1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use proximitybeacon1_beta1::Proximitybeacon;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Proximitybeacon::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.beacons().diagnostics_list("beaconName")
             .project_id("invidunt")
             .page_token("aliquyam")
             .page_size(-73)
             .alert_filter("Lorem")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AdvertisedId
Defines a unique identifier of a beacon as broadcast by the device.

AttachmentInfo
A subset of attachment information served via the beaconinfo.getforobserved method, used when your users encounter your beacons.

Beacon
Details of a beacon device.

BeaconActivateCall
Activates a beacon. A beacon that is active will return information and attachment data when queried via beaconinfo.getforobserved. Calling this method on an already active beacon will do nothing (but will return a successful response code). Authenticate using an OAuth access token from a signed-in user with Is owner or Can edit permissions in the Google Developers Console project.

BeaconAttachment
Project-specific data associated with a beacon.

BeaconAttachmentBatchDeleteCall
Deletes multiple attachments on a given beacon. This operation is permanent and cannot be undone. You can optionally specify namespacedType to choose which attachments should be deleted. If you do not specify namespacedType, all your attachments on the given beacon will be deleted. You also may explicitly specify */* to delete all. Authenticate using an OAuth access token from a signed-in user with Is owner or Can edit permissions in the Google Developers Console project.

BeaconAttachmentCreateCall
Associates the given data with the specified beacon. Attachment data must contain two parts: - A namespaced type. - The actual attachment data itself. The namespaced type consists of two parts, the namespace and the type. The namespace must be one of the values returned by the namespaces endpoint, while the type can be a string of any characters except for the forward slash (/) up to 100 characters in length. Attachment data can be up to 1024 bytes long. Authenticate using an OAuth access token from a signed-in user with Is owner or Can edit permissions in the Google Developers Console project.

BeaconAttachmentDeleteCall
Deletes the specified attachment for the given beacon. Each attachment has a unique attachment name (attachmentName) which is returned when you fetch the attachment data via this API. You specify this with the delete request to control which attachment is removed. This operation cannot be undone. Authenticate using an OAuth access token from a signed-in user with Is owner or Can edit permissions in the Google Developers Console project.

BeaconAttachmentListCall
Returns the attachments for the specified beacon that match the specified namespaced-type pattern. To control which namespaced types are returned, you add the namespacedType query parameter to the request. You must either use */*, to return all attachments, or the namespace must be one of the ones returned from the namespaces endpoint. Authenticate using an OAuth access token from a signed-in user with viewer, Is owner or Can edit permissions in the Google Developers Console project.

BeaconDeactivateCall
Deactivates a beacon. Once deactivated, the API will not return information nor attachment data for the beacon when queried via beaconinfo.getforobserved. Calling this method on an already inactive beacon will do nothing (but will return a successful response code). Authenticate using an OAuth access token from a signed-in user with Is owner or Can edit permissions in the Google Developers Console project.

BeaconDecommissionCall
Decommissions the specified beacon in the service. This beacon will no longer be returned from beaconinfo.getforobserved. This operation is permanent -- you will not be able to re-register a beacon with this ID again. Authenticate using an OAuth access token from a signed-in user with Is owner or Can edit permissions in the Google Developers Console project.

BeaconDiagnosticListCall
List the diagnostics for a single beacon. You can also list diagnostics for all the beacons owned by your Google Developers Console project by using the beacon name beacons/-. Authenticate using an OAuth access token from a signed-in user with viewer, Is owner or Can edit permissions in the Google Developers Console project.

BeaconGetCall
Returns detailed information about the specified beacon. Authenticate using an OAuth access token from a signed-in user with viewer, Is owner or Can edit permissions in the Google Developers Console project. Requests may supply an Eddystone-EID beacon name in the form: beacons/4!beaconId where the beaconId is the base16 ephemeral ID broadcast by the beacon. The returned Beacon object will contain the beacon's stable Eddystone-UID. Clients not authorized to resolve the beacon's ephemeral Eddystone-EID broadcast will receive an error.

BeaconInfo
A subset of beacon information served via the beaconinfo.getforobserved method, which you call when users of your app encounter your beacons.

BeaconListCall
Searches the beacon registry for beacons that match the given search criteria. Only those beacons that the client has permission to list will be returned. Authenticate using an OAuth access token from a signed-in user with viewer, Is owner or Can edit permissions in the Google Developers Console project.

BeaconMethods
A builder providing access to all methods supported on beacon resources. It is not used directly, but through the Proximitybeacon hub.

BeaconRegisterCall
Registers a previously unregistered beacon given its advertisedId. These IDs are unique within the system. An ID can be registered only once. Authenticate using an OAuth access token from a signed-in user with Is owner or Can edit permissions in the Google Developers Console project.

BeaconUpdateCall
Updates the information about the specified beacon. Any field that you do not populate in the submitted beacon will be permanently erased, so you should follow the "read, modify, write" pattern to avoid inadvertently destroying data. Changes to the beacon status via this method will be silently ignored. To update beacon status, use the separate methods on this API for activation, deactivation, and decommissioning. Authenticate using an OAuth access token from a signed-in user with Is owner or Can edit permissions in the Google Developers Console project.

BeaconinfoGetforobservedCall
Given one or more beacon observations, returns any beacon information and attachments accessible to your application. Authorize by using the API key for the application.

BeaconinfoMethods
A builder providing access to all methods supported on beaconinfo resources. It is not used directly, but through the Proximitybeacon hub.

Date
Represents a whole calendar date, e.g. date of birth. The time of day and time zone are either specified elsewhere or are not significant. The date is relative to the Proleptic Gregorian Calendar. The day may be 0 to represent a year and month where the day is not significant, e.g. credit card expiration date. The year may be 0 to represent a month and day independent of year, e.g. anniversary date. Related types are google.type.TimeOfDay and google.protobuf.Timestamp.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DeleteAttachmentsResponse
Response for a request to delete attachments.

Diagnostics
Diagnostics for a single beacon.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

EphemeralIdRegistration
Write-only registration parameters for beacons using Eddystone-EID format. Two ways of securely registering an Eddystone-EID beacon with the service are supported: 1. Perform an ECDH key exchange via this API, including a previous call to GET /v1beta1/eidparams. In this case the fields beacon_ecdh_public_key and service_ecdh_public_key should be populated and beacon_identity_key should not be populated. This method ensures that only the two parties in the ECDH key exchange can compute the identity key, which becomes a secret between them. 2. Derive or obtain the beacon's identity key via other secure means (perhaps an ECDH key exchange between the beacon and a mobile device or any other secure method), and then submit the resulting identity key to the service. In this case beacon_identity_key field should be populated, and neither of beacon_ecdh_public_key nor service_ecdh_public_key fields should be. The security of this method depends on how securely the parties involved (in particular the bluetooth client) handle the identity key, and obviously on how securely the identity key was generated. See the Eddystone specification at GitHub.

EphemeralIdRegistrationParams
Information a client needs to provision and register beacons that broadcast Eddystone-EID format beacon IDs, using Elliptic curve Diffie-Hellman key exchange. See the Eddystone specification at GitHub.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

GetInfoForObservedBeaconsRequest
Request for beacon and attachment information about beacons that a mobile client has encountered "in the wild".

GetInfoForObservedBeaconsResponse
Information about the requested beacons, optionally including attachment data.

IndoorLevel
Indoor level, a human-readable string as returned by Google Maps APIs, useful to indicate which floor of a building a beacon is located on.

LatLng
An object representing a latitude/longitude pair. This is expressed as a pair of doubles representing degrees latitude and degrees longitude. Unless specified otherwise, this must conform to the WGS84 standard. Values must be within normalized ranges. Example of normalization code in Python: def NormalizeLongitude(longitude): """Wraps decimal degrees longitude to [-180.0, 180.0].""" q, r = divmod(longitude, 360.0) if r > 180.0 or (r == 180.0 and q <= -1.0): return r - 360.0 return r def NormalizeLatLng(latitude, longitude): """Wraps decimal degrees latitude and longitude to [-90.0, 90.0] and [-180.0, 180.0], respectively.""" r = latitude % 360.0 if r = 270.0: return r - 360, NormalizeLongitude(longitude) else: return 180 - r, NormalizeLongitude(longitude + 180.0) assert 180.0 == NormalizeLongitude(180.0) assert -180.0 == NormalizeLongitude(-180.0) assert -179.0 == NormalizeLongitude(181.0) assert (0.0, 0.0) == NormalizeLatLng(360.0, 0.0) assert (0.0, 0.0) == NormalizeLatLng(-360.0, 0.0) assert (85.0, 180.0) == NormalizeLatLng(95.0, 0.0) assert (-85.0, -170.0) == NormalizeLatLng(-95.0, 10.0) assert (90.0, 10.0) == NormalizeLatLng(90.0, 10.0) assert (-90.0, -10.0) == NormalizeLatLng(-90.0, -10.0) assert (0.0, -170.0) == NormalizeLatLng(-180.0, 10.0) assert (0.0, -170.0) == NormalizeLatLng(180.0, 10.0) assert (-90.0, 10.0) == NormalizeLatLng(270.0, 10.0) assert (90.0, 10.0) == NormalizeLatLng(-270.0, 10.0)

ListBeaconAttachmentsResponse
Response to ListBeaconAttachments that contains the requested attachments.

ListBeaconsResponse
Response that contains list beacon results and pagination help.

ListDiagnosticsResponse
Response that contains the requested diagnostics.

ListNamespacesResponse
Response to ListNamespacesRequest that contains all the project's namespaces.

MethodGetEidparamCall
Gets the Proximity Beacon API's current public key and associated parameters used to initiate the Diffie-Hellman key exchange required to register a beacon that broadcasts the Eddystone-EID format. This key changes periodically; clients may cache it and re-use the same public key to provision and register multiple beacons. However, clients should be prepared to refresh this key when they encounter an error registering an Eddystone-EID beacon.

MethodInfo
Contains information about an API request.

MethodMethods
A builder providing access to all free methods, which are not associated with a particular resource. It is not used directly, but through the Proximitybeacon hub.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Namespace
An attachment namespace defines read and write access for all the attachments created under it. Each namespace is globally unique, and owned by one project which is the only project that can create attachments under it.

NamespaceListCall
Lists all attachment namespaces owned by your Google Developers Console project. Attachment data associated with a beacon must include a namespaced type, and the namespace must be owned by your project. Authenticate using an OAuth access token from a signed-in user with viewer, Is owner or Can edit permissions in the Google Developers Console project.

NamespaceMethods
A builder providing access to all methods supported on namespace resources. It is not used directly, but through the Proximitybeacon hub.

NamespaceUpdateCall
Updates the information about the specified namespace. Only the namespace visibility can be updated.

Observation
Represents one beacon observed once.

Proximitybeacon
Central instance to access all Proximitybeacon related resource activities

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.