Struct google_proximitybeacon1_beta1::AdvertisedId

pub struct AdvertisedId {
    pub type_: Option<String>,
    pub id: Option<String>,
}

Defines a unique identifier of a beacon as broadcast by the device.

This type is not used in any activity, and only used as part of another schema.

Fields
type_: Option<String>

Specifies the identifier type. Required.

id: Option<String>

The actual beacon identifier, as broadcast by the beacon hardware. Must be base64 encoded in HTTP requests, and will be so encoded (with padding) in responses. The base64 encoding should be of the binary byte-stream and not any textual (such as hex) representation thereof. Required.