Struct google_proximitybeacon1_beta1::Beacon

pub struct Beacon {
    pub status: Option<String>,
    pub description: Option<String>,
    pub indoor_level: Option<IndoorLevel>,
    pub lat_lng: Option<LatLng>,
    pub place_id: Option<String>,
    pub ephemeral_id_registration: Option<EphemeralIdRegistration>,
    pub advertised_id: Option<AdvertisedId>,
    pub beacon_name: Option<String>,
    pub expected_stability: Option<String>,
    pub properties: Option<HashMap<String, String>>,
    pub provisioning_key: Option<String>,
}

Details of a beacon device.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

attachments list beacons (none)

get beacons (response)

attachments create beacons (none)

decommission beacons (none)

activate beacons (none)

list beacons (none)

update beacons (request|response)

attachments delete beacons (none)

deactivate beacons (none)

register beacons (request|response)

diagnostics list beacons (none)

attachments batch delete beacons (none)

Fields
status: Option<String>

Current status of the beacon. Required.

description: Option<String>

Free text used to identify and describe the beacon. Maximum length 140 characters. Optional.

indoor_level: Option<IndoorLevel>

The indoor level information for this beacon, if known. As returned by the Google Maps API. Optional.

lat_lng: Option<LatLng>

The location of the beacon, expressed as a latitude and longitude pair. This location is given when the beacon is registered or updated. It does not necessarily indicate the actual current location of the beacon. Optional.

place_id: Option<String>

The Google Places API Place ID of the place where the beacon is deployed. This is given when the beacon is registered or updated, not automatically detected in any way. Optional.

ephemeral_id_registration: Option<EphemeralIdRegistration>

Write-only registration parameters for beacons using Eddystone-EID (remotely resolved ephemeral ID) format. This information will not be populated in API responses. When submitting this data, the advertised_id field must contain an ID of type Eddystone-UID. Any other ID type will result in an error.

advertised_id: Option<AdvertisedId>

The identifier of a beacon as advertised by it. This field must be populated when registering. It may be empty when updating a beacon record because it is ignored in updates. When registering a beacon that broadcasts Eddystone-EID, this field should contain a "stable" Eddystone-UID that identifies the beacon and links it to its attachments. The stable Eddystone-UID is only used for administering the beacon.

beacon_name: Option<String>

Resource name of this beacon. A beacon name has the format "beacons/N!beaconId" where the beaconId is the base16 ID broadcast by the beacon and N is a code for the beacon's type. Possible values are 3 for Eddystone, 1 for iBeacon, or 5 for AltBeacon. This field must be left empty when registering. After reading a beacon, clients can use the name for future operations.

expected_stability: Option<String>

Expected location stability. This is set when the beacon is registered or updated, not automatically detected in any way. Optional.

properties: Option<HashMap<String, String>>

Properties of the beacon device, for example battery type or firmware version. Optional.

provisioning_key: Option<String>

Some beacons may require a user to provide an authorization key before changing any of its configuration (e.g. broadcast frames, transmit power). This field provides a place to store and control access to that key. This field is populated in responses to GET /v1beta1/beacons/3!beaconId from users with write access to the given beacon. That is to say: If the user is authorized to write the beacon's confidential data in the service, the service considers them authorized to configure the beacon. Note that this key grants nothing on the service, only on the beacon itself.