Struct google_proximitybeacon1_beta1::BeaconAttachment

pub struct BeaconAttachment {
    pub attachment_name: Option<String>,
    pub data: Option<String>,
    pub namespaced_type: Option<String>,
}

Project-specific data associated with a beacon.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

attachments create beacons (request|response)

Fields
attachment_name: Option<String>

Resource name of this attachment. Attachment names have the format: beacons/beacon_id/attachments/attachment_id. Leave this empty on creation.

data: Option<String>

An opaque data container for client-provided data. Must be base64 encoded in HTTP requests, and will be so encoded (with padding) in responses. Required.

namespaced_type: Option<String>

Specifies what kind of attachment this is. Tells a client how to interpret the data field. Format is namespace/type. Namespace provides type separation between clients. Type describes the type of data, for use by the client when parsing the data field. Required.