Struct google_proximitybeacon1_beta1::BeaconListCall

pub struct BeaconListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Searches the beacon registry for beacons that match the given search criteria. Only those beacons that the client has permission to list will be returned.

A builder for the list method supported by a beacon resource. It is not used directly, but through a BeaconMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.beacons().list()
             .q("nonumy")
             .page_token("dolores")
             .page_size(-61)
             .doit();

Methods

impl<'a, C, A> BeaconListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, ListBeaconsResponse)>

Perform the operation you have build so far.

fn q(self, new_value: &str) -> BeaconListCall<'a, C, A>

Filter query string that supports the following field filters: * description:"" For example: description:"Room 3" Returns beacons whose description matches tokens in the string "Room 3" (not necessarily that exact string). The string must be double-quoted. * status: For example: status:active Returns beacons whose status matches the given value. Values must be one of the Beacon.Status enum values (case insensitive). Accepts multiple filters which will be combined with OR logic. * stability: For example: stability:mobile Returns beacons whose expected stability matches the given value. Values must be one of the Beacon.Stability enum values (case insensitive). Accepts multiple filters which will be combined with OR logic. * place_id:"" For example: place_id:"ChIJVSZzVR8FdkgRXGmmm6SslKw=" Returns beacons explicitly registered at the given place, expressed as a Place ID obtained from Google Places API. Does not match places inside the given place. Does not consider the beacon's actual location (which may be different from its registered place). Accepts multiple filters that will be combined with OR logic. The place ID must be double-quoted. * registration_time[|=] For example: registration_time>=1433116800 Returns beacons whose registration time matches the given filter. Supports the operators: , =. Timestamp must be expressed as an integer number of seconds since midnight January 1, 1970 UTC. Accepts at most two filters that will be combined with AND logic, to support "between" semantics. If more than two are supplied, the latter ones are ignored. * lat: lng: radius: For example: lat:51.1232343 lng:-1.093852 radius:1000 Returns beacons whose registered location is within the given circle. When any of these fields are given, all are required. Latitude and longitude must be decimal degrees between -90.0 and 90.0 and between -180.0 and 180.0 respectively. Radius must be an integer number of meters less than 1,000,000 (1000 km). * property:"=" For example: property:"battery-type=CR2032" Returns beacons which have a property of the given name and value. Supports multiple filters which will be combined with OR logic. The entire name=value string must be double-quoted as one string. * attachment_type:"" For example: attachment_type:"my-namespace/my-type" Returns beacons having at least one attachment of the given namespaced type. Supports "any within this namespace" via the partial wildcard syntax: "my-namespace/*". Supports multiple filters which will be combined with OR logic. The string must be double-quoted. Multiple filters on the same field are combined with OR logic (except registration_time which is combined with AND logic). Multiple filters on different fields are combined with AND logic. Filters should be separated by spaces. As with any HTTP query string parameter, the whole filter expression must be URL-encoded. Example REST request: GET /v1beta1/beacons?q=status:active%20lat:51.123%20lng:-1.095%20radius:1000

Sets the q query property to the given value.

fn page_token(self, new_value: &str) -> BeaconListCall<'a, C, A>

A pagination token obtained from a previous request to list beacons.

Sets the page token query property to the given value.

fn page_size(self, new_value: i32) -> BeaconListCall<'a, C, A>

The maximum number of records to return for this request, up to a server-defined upper limit.

Sets the page size query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> BeaconListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> BeaconListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

bearer_token (query-string) - OAuth bearer token.

pp (query-boolean) - Pretty-print response.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

upload_protocol (query-string) - Upload protocol for media (e.g. "raw", "multipart").

access_token (query-string) - OAuth access token.

fields (query-string) - Selector specifying which fields to include in a partial response.

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.

callback (query-string) - JSONP

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

uploadType (query-string) - Legacy upload protocol for media (e.g. "media", "multipart").

alt (query-string) - Data format for response.

$.xgafv (query-string) - V1 error format.