Struct google_proximitybeacon1_beta1::ListDiagnosticsResponse

pub struct ListDiagnosticsResponse {
    pub next_page_token: Option<String>,
    pub diagnostics: Option<Vec<Diagnostics>>,
}

Response that contains the requested diagnostics.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

diagnostics list beacons (response)

Fields
next_page_token: Option<String>

Token that can be used for pagination. Returned only if the request matches more beacons than can be returned in this response.

diagnostics: Option<Vec<Diagnostics>>

The diagnostics matching the given request.