Struct google_proximitybeacon1_beta1::NamespaceMethods

pub struct NamespaceMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on namespace resources. It is not used directly, but through the Proximitybeacon hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_proximitybeacon1_beta1 as proximitybeacon1_beta1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use proximitybeacon1_beta1::Proximitybeacon;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Proximitybeacon::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `list(...)` and `update(...)`
// to build up your call.
let rb = hub.namespaces();

Methods

impl<'a, C, A> NamespaceMethods<'a, C, A>

fn list(&self) -> NamespaceListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all attachment namespaces owned by your Google Developers Console project. Attachment data associated with a beacon must include a namespaced type, and the namespace must be owned by your project. Authenticate using an OAuth access token from a signed-in user with viewer, Is owner or Can edit permissions in the Google Developers Console project.

fn update(&self, request: Namespace, namespace_name: &str) -> NamespaceUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the information about the specified namespace. Only the namespace visibility can be updated.

Arguments

request - No description provided.

namespaceName - Resource name of this namespace. Namespaces names have the format: namespaces/namespace.